[FEAT] DoubleCaretIcon 추가, caret 공용 값 정리 | Added DoubleCaretIcon and organized shared caret values TALEARNT-373
